LayoutTests/fast/dom/xmlserializer-serialize-to-string-exception-expected.txt

This tests XMLSerializer.serializeToString() throwing exception when node value is invalid and passing otherwise.

1. Verifying XMLSerializer.serializeToString() should THROW exception with node value = null
Exception thrown = [TypeError: Type error]
PASS

2. Verifying XMLSerializer.serializeToString() should THROW exception with node value = undefined
Exception thrown = [TypeError: Type error]
PASS

3. Verifying XMLSerializer.serializeToString() should THROW exception with node value = <html><title>Hello World</title></html>
Exception thrown = [TypeError: Type error]
PASS

4. Verifying XMLSerializer.serializeToString() should THROW exception with node value = [object HTMLCollection]
Exception thrown = [TypeError: Type error]
PASS

5. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLDocument]
PASS

6. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLHtmlElement]
PASS

7. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLHtmlElement]
PASS

8. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLDivElement]
PASS

9. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLHeadingElement]
PASS

10. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLUnknownElement]
PASS

11. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLDocument]
PASS

12. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object HTMLHtmlElement]
PASS

13. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object XMLDocument]
PASS

14. Verifying XMLSerializer.serializeToString() should NOT-THROW exception with node value = [object Element]
PASS
